+++ pi-provider-sub2api/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Unreleased
+## 0.1.27 - 2026-09-11
+
+- Synchronize the package version with `@indexyz/pi-custom-provider` for coordinated workspace releases.
 
 ## 0.1.26 - 2026-09-06